Rename IndentedPrinter to CodeBuffer and add Examples.
CodeBuffer is a name that better reflects the functionality of the tool.
Also changed is the namespace that users will access the tool on. You
will now use simplegen::CodeBuffer to access the libraries
functionality.

The added examples in the docs should make it a little clearer how to
use the library, however I think this still requires a decent bit of
work.
