+++ Cargo.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 keywords = ["codegen"]
 categories = ["template-engine"]
 license = "GPL-2.0-only"
-version = "0.1.2"
+version = "0.2.0"
 edition = "2018"
 
 [badges]
