+++ src/lib.rs
@@ -1,1 +1,3 @@
-pub mod printer;
+mod code_buffer;
+
+pub use code_buffer::CodeBuffer;
